requirements.txt
apscheduler
psutil
requests
matplotlib
Pillow
colorama
pytesseract
baidu-aip
apscheduler
chardet